+++ tests/testthat/test-first_isolate.R
@@ -10,7 +10,7 @@ test_that("first isolates work", {
                     col_mo = "mo",
                     info = TRUE),
       na.rm = TRUE),
-    1331)
+    1330)
 
   # septic_patients contains 1426 out of 2000 first *weighted* isolates
   expect_equal(
@@ -24,7 +24,7 @@ test_that("first isolates work", {
                       type = "keyantibiotics",
                       info = TRUE),
         na.rm = TRUE)),
-    1426)
+    1425)
   # and 1449 when not ignoring I
   expect_equal(
     suppressWarnings(
@@ -38,7 +38,7 @@ test_that("first isolates work", {
                       type = "keyantibiotics",
                       info = TRUE),
         na.rm = TRUE)),
-    1449)
+    1448)
   # and 1430 when using points
   expect_equal(
     suppressWarnings(
@@ -64,7 +64,7 @@ test_that("first isolates work", {
                     info = TRUE,
                     icu_exclude = TRUE),
       na.rm = TRUE),
-    1176)
+    1175)
 
   # set 1500 random observations to be of specimen type 'Urine'
   random_rows <- sample(x = 1:2000, size = 1500, replace = FALSE)