+++ NEWS.md
@@ -10,14 +10,16 @@
   * Column names of datasets `microorganisms` and `septic_patients`
   * All old syntaxes will still work with this version, but will throw warnings
 * Functions `as.atc` and `is.atc` to transform/look up antibiotic ATC codes as defined by the WHO. The existing function `guess_atc` is now an alias of `as.atc`.
-* Aliases for existing function `mo_property`: `mo_family`, `mo_genus`, `mo_species`, `mo_subspecies`, `mo_fullname`, `mo_shortname`, `mo_aerobic`, `mo_type` and `mo_gramstain`. The last two functions have a `language` parameter, with support for Spanish, German and Dutch:
+* Aliases for existing function `mo_property`: `mo_family`, `mo_genus`, `mo_species`, `mo_subspecies`, `mo_fullname`, `mo_shortname`, `mo_aerobic`, `mo_type` and `mo_gramstain`. They also come with support for German, Dutch, Spanish and Portuguese, and it defaults to the systems locale:
   ```r
   mo_gramstain("E. coli")
   # [1] "Negative rods"
   mo_gramstain("E. coli", language = "de") # "de" = Deutsch / German
-  # [1] "Negative Staebchen"
+  # [1] "Negative Stäbchen"
   mo_gramstain("E. coli", language = "es") # "es" = Español / Spanish
   # [1] "Bacilos negativos"
+  mo_fullname("S. group A") # when run on a on a Portuguese system
+  # [1] "Streptococcus grupo A"
   ```
 * Function `ab_property` and its aliases: `ab_official`, `ab_tradenames`, `ab_certe`, `ab_umcg`, `ab_official_nl` and `ab_trivial_nl`
 * Introduction to AMR as a vignette
@@ -34,6 +36,7 @@
   ab_atc(c("Bactroban", "Amoxil", "Zithromax", "Floxapen"))
   # [1] "R01AX06" "J01CA04" "J01FA10" "J01CF05"
   ```
+* For `first_isolate`, rows will be ignored when there's no species available
 * Function `ratio` is now deprecated and will be removed in a future release, as it is not really the scope of this package
 * Fix for `as.mic` for values ending in zeroes after a real number
 * Tremendous speed improvement for `as.bactid` (now `as.mo`)